[Blink-19382226]Fix shuffle merge integer overflow bug.
Summary:
[Blink-19382226]Fix shuffle merge integer overflow bug.

fix T19382226
ref T19382226

Reviewers: 云骞

Reviewed By: 云骞

Subscribers: P710480

Differential Revision: https://aone.alibaba-inc.com/code/D855380

+++ flink-runtime/src/main/java/org/apache/flink/runtime/io/network/partition/external/writer/ConcatPartitionedFileMerger.java
@@ -202,8 +202,8 @@ public class ConcatPartitionedFileMerger<T> implements SortedDataFileMerger<Tupl
 				heap.adjustTop();
 			}
 
-			// now read the specific counts of buffers
-			int readLength = 0;
+			// now read the specific length of data
+			long readLength = 0;
 			while (readLength < partitionIndex.getLength()) {
 				Buffer buffer = headStream.getReader().getNextBufferBlocking();
 				readLength += buffer.getSize();
